change volume to float

+++ QT/sounddisplay.cpp
@@ -124,7 +124,7 @@ void SoundDisplay::volumeChanged(int changedVolume)
 {
     volume = changedVolume;
     volumeLabel->setText(QString::fromStdString("Volume: " + std::to_string(volume)));
-    double scaledVolume = (double)(volume)/100;
+    float scaledVolume = (float)(volume)/100;
     this->soundFile.changeVolume(scaledVolume);
     // Call other function which actually changes volume
     //change the volume
